Refactor GUI tests
